output/httpd/IcyMetaDataServer: cast length to unsigned
Fixes another buffer overflow: if the stream has a very long title or
URL, resulting in a metadata string of more than 2 kB, icy_string[0]
is a negative value, which gets casted to size_t - ouch!

 https://bugs.musicpd.org/view.php?id=4652

+++ src/output/plugins/httpd/IcyMetaDataServer.cxx
@@ -113,5 +113,5 @@ icy_server_metadata_page(const Tag &tag, const TagType *types)
 	if (icy_string.IsNull())
 		return nullptr;
 
-	return Page::Copy(icy_string.c_str(), (icy_string[0] * 16) + 1);
+	return Page::Copy(icy_string.c_str(), uint8_t(icy_string[0]) * 16 + 1);
 }
